修改core包版本为3.0.1

+++ web/service/src/main/java/common/com/emenu/common/dto/dish/DishDto.java
@@ -44,7 +44,7 @@ public class DishDto {
     private Integer saleType;
 
     // 折扣
-    private Integer discount;
+    private Float discount;
 
     // 售价
     private BigDecimal salePrice;
@@ -162,11 +162,11 @@ public class DishDto {
         this.saleType = saleType;
     }
 
-    public Integer getDiscount() {
+    public Float getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Integer discount) {
+    public void setDiscount(Float discount) {
         this.discount = discount;
     }
